requirements.txt
# Core Dependencies
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# LLM
zhipuai>=4.0.0

# Notion
pynotion>=0.3.0

# Scheduler
apscheduler>=3.10.0

# Utilities
python-dateutil>=2.8.0
requests>=2.31.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
